requirements.txt
appdirs==1.4.4
asgiref==3.4.1
basicsynbio==0.5.0
biopython==1.79
black==21.7b0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
Deprecated==1.2.12
fastapi==0.68.0
h11==0.12.0
idna==3.2
isodate==0.6.0
lxml==4.6.3
mypy-extensions==0.4.3
numpy==1.21.2
packaging==21.0
pandas==1.3.2
pathspec==0.9.0
Pillow==8.3.1
platemap==0.0.10
pydantic==1.8.2
pyparsing==2.4.7
python-dateutil==2.8.2
python-Levenshtein==0.12.2
pytz==2021.1
rdflib==6.0.0
regex==2021.8.3
reportlab==3.6.1
requests==2.26.0
sbol2==1.2
six==1.16.0
starlette==0.14.2
tomli==1.2.1
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.15.0
wrapt==1.12.1
